requirements.txt
ctransformers
sentence-transformers
pinecone
langchain
python-dotenv
langchain-community
pypdf
langchain-text-splitters
langchain-pinecone
langchain-huggingface
flask
-e .